requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
django-crontab==0.7.1
djangorestframework==3.14.0
django-celery-beat==2.5.0
tzdata
celery==5.2.7
google_api_python_client==2.84.0
flower==1.2.0
numpy==1.23.5
pandas==1.5.3
Pyrebase4==4.6.0
requests==2.28.2
scikit-learn==1.2.2
redis==4.5.4

